refactor: Auslagern der Methoden in neue Klasse, Anpassen der Hochrechnung
